chore(examples): move `app` directories
