+++ examples/with-redux/src/features/counter/Counter.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { useAppSelector, useAppDispatch } from '../../app/hooks'
+import { useAppSelector, useAppDispatch } from '../../hooks'
 import {
   decrement,
   increment,
