+++ examples/with-redux/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import '../styles/globals.css'
 import { Provider } from 'react-redux'
 import type { AppProps } from 'next/app'
 
-import store from '../app/store'
+import store from '../store'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
