+++ examples/with-redux/src/store.ts
